requirements.txt
altgraph==0.16.1
appdirs==1.4.3
astroid==2.1.0
attrs==18.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
black==18.9b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
macholib==1.11
mccabe==0.6.1
modulegraph==0.17
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==5.4.1
py2app==0.19
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.2.2
PyOpenGL==3.1.0
PyQt5==5.11.2
PyQt5-sip==4.19.12
python-dotenv==0.15.0
pytz==2020.4
requests==2.21.0
six==1.12.0
soupsieve==2.0.1
toml==0.10.0
twilio==6.49.0
urllib3==1.24.1
wrapt==1.11.1
wxPython==4.0.4
